ULay.dp2px方法参数和返回值稍作调整 新增ItemBar底栏控件

+++ libs-android/src/main/java/candyenk/android/widget/DialogBottom.java
@@ -100,7 +100,7 @@ public class DialogBottom extends BottomSheetDialog {
         super.dismiss();
     }
 
-    protected double dp2px(double dp) {
+    protected float dp2px(float dp) {
         return ULay.dp2px(context, dp);
     }
     /**********************************************************************************************/
